+++ test/media-segment-request.test.js
@@ -30,6 +30,19 @@ QUnit.module('Media Segment Request', {
         this.listeners = this.listeners.filter((fn)=>fn !== listener);
       }
     };
+    this.mockCaptionParser = {
+      initialized: false,
+      parsed: false,
+      isInitialized() {
+        return this.initialized;
+      },
+      init() {
+        this.initialized = true;
+      },
+      parse(segment, videoTrackIds, timescales) {
+        this.parsed = true;
+      }
+    };
     this.xhrOptions = {
       timeout: 1000
     };
@@ -50,6 +63,7 @@ QUnit.test('cancels outstanding segment request on abort', function(assert) {
     xhr: this.xhr,
     xhrOptions: this.xhrOptions,
     decryptionWorker: this.noop,
+    captionParser: this.noop,
     segment: { resolvedUri: '0-test.ts' },
     progressFn: this.noop,
     doneFn: (error, segmentData) => {
@@ -81,6 +95,7 @@ QUnit.test('cancels outstanding key requests on abort', function(assert) {
     xhr: this.xhr,
     xhrOptions: this.xhrOptions,
     decryptionWorker: this.noop,
+    captionParser: this.noop,
     segment: {
       resolvedUri: '0-test.ts',
       key: {
@@ -120,6 +135,7 @@ QUnit.test('cancels outstanding key requests on failure', function(assert) {
     xhr: this.xhr,
     xhrOptions: this.xhrOptions,
     decryptionWorker: this.noop,
+    captionParser: this.noop,
     segment: {
       resolvedUri: '0-test.ts',
       key: {
@@ -156,6 +172,7 @@ QUnit.test('cancels outstanding key requests on timeout', function(assert) {
     xhr: this.xhr,
     xhrOptions: this.xhrOptions,
     decryptionWorker: this.noop,
+    captionParser: this.noop,
     segment: {
       resolvedUri: '0-test.ts',
       key: {
@@ -203,6 +220,7 @@ QUnit.test('the key response is converted to the correct format', function(asser
     xhr: this.xhr,
     xhrOptions: this.xhrOptions,
     decryptionWorker: this.mockDecrypter,
+    captionParser: this.noop,
     segment: {
       resolvedUri: '0-test.ts',
       key: {
@@ -244,6 +262,7 @@ QUnit.test('segment with key has bytes decrypted', function(assert) {
     xhr: this.xhr,
     xhrOptions: this.xhrOptions,
     decryptionWorker: this.realDecrypter,
+    captionParser: this.noop,
     segment: {
       resolvedUri: '0-test.ts',
       key: {
@@ -290,6 +309,7 @@ function(assert) {
     xhr: this.xhr,
     xhrOptions: this.xhrOptions,
     decryptionWorker: this.realDecrypter,
+    captionParser: this.mockCaptionParser,
     segment: {
       resolvedUri: '0-test.ts',
       key: {
@@ -337,4 +357,135 @@ function(assert) {
 
   // Allow the decrypter to decrypt
   this.clock.tick(100);
+});
+
+QUnit.test('non-TS segment will get parsed for captions', function(assert) {
+  const done = assert.async();
+
+  mediaSegmentRequest(
+    this.xhr,
+    this.xhrOptions,
+    this.mockDecrypter,
+    this.mockCaptionParser,
+    {
+      resolvedUri: '0-test.m4s',
+      map: {
+        resolvedUri: '0-init.mp4'
+      }
+    },
+    this.noop,
+    (error, segmentData) => {
+      assert.notOk(error, 'there are no errors');
+      assert.ok(segmentData.map.bytes, 'init segment bytes in map');
+
+      // verify stats
+      assert.equal(segmentData.stats.bytesReceived, 8, '8 bytes');
+      // verify cached map
+      assert.ok(segmentData.map.timescales, 'looked for timescales');
+      assert.ok(segmentData.map.videoTrackIds, 'looked for videoTrackIds');
+      // verify the caption parser
+      assert.equal(this.mockCaptionParser.parsed, true, 'tried to parse captions');
+      done();
+    });
+
+  assert.equal(this.requests.length, 2, 'there are two requests');
+
+  const initReq = this.requests.shift();
+  const segmentReq = this.requests.shift();
+
+  assert.equal(initReq.uri, '0-init.mp4', 'the first request is for the init segment');
+  assert.equal(segmentReq.uri, '0-test.m4s', 'the second request is for a segment');
+
+  initReq.response = new Uint32Array([0, 1, 2, 3]).buffer;
+  initReq.respond(200, null, '');
+  this.clock.tick(200);
+
+  segmentReq.response = new Uint8Array(8).buffer;
+  segmentReq.respond(200, null, '');
+  this.clock.tick(200);
+});
+
+QUnit.test('non-TS segment will get parsed for captions on next segment request if init is late', function(assert) {
+  const done = assert.async();
+  let initBytes;
+
+  mediaSegmentRequest(
+    this.xhr,
+    this.xhrOptions,
+    this.mockDecrypter,
+    this.mockCaptionParser,
+    {
+      resolvedUri: '0-test.m4s',
+      map: {
+        resolvedUri: '0-init.mp4'
+      }
+    },
+    this.noop,
+    (error, segmentData) => {
+      assert.notOk(error, 'there are no errors');
+      assert.ok(segmentData.map.bytes, 'init segment bytes in map');
+      initBytes = segmentData.map.bytes;
+
+      // verify stats
+      assert.equal(segmentData.stats.bytesReceived, 8, '8 bytes');
+      // verify cached map
+      assert.ok(segmentData.map.timescales, 'looked for timescales');
+      assert.ok(segmentData.map.videoTrackIds, 'looked for videoTrackIds');
+      // verify the caption parser
+      assert.equal(this.mockCaptionParser.parsed, false, 'tried to parse captions');
+    });
+
+  assert.equal(this.requests.length, 2, 'there are two requests');
+
+  let initReq = this.requests.shift();
+  let segmentReq = this.requests.shift();
+
+  assert.equal(initReq.uri, '0-init.mp4', 'the first request is for the init segment');
+  assert.equal(segmentReq.uri, '0-test.m4s', 'the second request is for a segment');
+
+  segmentReq.response = new Uint8Array(8).buffer;
+  segmentReq.respond(200, null, '');
+  this.clock.tick(200);
+
+  initReq.response = new Uint32Array([0, 1, 2, 3]).buffer;
+  initReq.respond(200, null, '');
+  this.clock.tick(200);
+
+  mediaSegmentRequest(
+    this.xhr,
+    this.xhrOptions,
+    this.mockDecrypter,
+    this.mockCaptionParser,
+    {
+      resolvedUri: '1-test.m4s',
+      map: {
+        resolvedUri: '0-init.mp4',
+        bytes: initBytes,
+        timescales: {},
+        videoTrackIds: [1]
+      }
+    },
+    this.noop,
+    (error, segmentData) => {
+      assert.notOk(error, 'there are no errors');
+      assert.ok(segmentData.map.bytes, 'init segment bytes in map');
+
+      // verify stats
+      assert.equal(segmentData.stats.bytesReceived, 8, '8 bytes');
+      // verify cached map
+      assert.ok(segmentData.map.timescales, 'looked for timescales');
+      assert.ok(segmentData.map.videoTrackIds, 'looked for videoTrackIds');
+      // verify the caption parser
+      assert.equal(this.mockCaptionParser.parsed, true, 'tried to parse captions');
+      done();
+    });
+
+  assert.equal(this.requests.length, 1, 'there is one request');
+
+  segmentReq = this.requests.shift();
+  assert.equal(segmentReq.uri, '1-test.m4s', 'the next request is for a segment');
+
+  segmentReq.response = new Uint8Array(8).buffer;
+  segmentReq.respond(200, null, '');
+  this.clock.tick(200);
 });